requirements.txt
# FastAPI ותליות
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Streamlit - ממשק משתמש
streamlit==1.28.1

# עיבוד PDF
pypdf==4.0.1

# Vector Database
chromadb==0.4.21

# HTTP requests
requests>=2.31.0

# Environment variables
python-dotenv==1.0.0

# Type hints ו-validation
pydantic==2.5.0

# בעתיד - embedding מקצועי
# sentence-transformers>=2.2.2

# Logging ו-monitoring
python-json-logger==2.0.7

# Security
cryptography>=41.0.0

